+++ trunk/SHM/src/main/java/ua/com/manometer/service/invoice/ShipmentMediatorService.java
@@ -10,6 +10,6 @@ public interface ShipmentMediatorService {
 
 	public List<ShipmentMediator> listShipmentMediator();
 
-	public void removeShipmentMediator(Long id);
+	public void removeShipmentMediator(Integer id);
 
 }
