+++ trunk/SHM/src/main/java/ua/com/manometer/service/invoice/ShipmentService.java
@@ -10,6 +10,6 @@ public interface ShipmentService {
 
 	public List<Shipment> listShipment();
 
-	public void removeShipment(Long id);
+	public void removeShipment(Integer id);
 
 }
